AK: Move heavyweight fuzzy matching to own compilation unit
